+++ apps/@examples/card/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @examples/card
 
+## 0.0.32
+
+### Patch Changes
+
+- Updated dependencies [a4209f4df]
+  - @dopt/react@5.0.2
+  - @dopt/react-card@10.0.2
+
 ## 0.0.31
 
 ### Patch Changes