+++ apps/@examples/help-hub/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @examples/help-hub
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [a4209f4df]
+  - @dopt/react@5.0.2
+  - @dopt/react-help-hub@0.1.1
+
 ## 0.0.18
 
 ### Patch Changes