+++ components/@dopt/semantic-data-layer/modal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dopt/semantic-data-layer-modal
 
+## 0.2.3
+
+### Patch Changes
+
+- @dopt/block-api-types@1.3.4
+
 ## 0.2.2
 
 ### Patch Changes